Added filter and summary programs
The program filterFoci ensures the intensity values within the 7x7
region surrounding the brightest pixel is skewed to the right. The
program filterPosition ensures the kinetochore foci are between the
min/max X and Y positions of the spindle pole body foci with a
parameter to add a margin. The noArrayOutliers program simply user
isoutliter function and a while loop to remove all outliers from an
array. The spotHeightStructure program automates the construction of a
structure array containing spot height information.